extended measureText to take a width arg and support textWrap
- if called with a second arg specifying maximum width, the text will either be truncated at the last word boundary that fits or, if textWrap=true, the line will be broken
- in multi-line situations, the `actualBoundingBox*` values will describe the union of all the individual line bounds
- the TextMetrics has been extended to contain a `lines` attr which is a list of line-by-line measurements
- each `lines` element contains x/w/width/height bounds, the y-offset of its baseline (using the current textBaseline mode) and char indices for the portion of the string typeset on that line

+++ native/src/typography.rs
@@ -220,14 +220,14 @@ pub fn from_text_baseline(mode:Baseline) -> String{
   }.to_string()
 }
 
-pub fn get_baseline_offset(metrics: &FontMetrics, mode:Baseline) -> f64 {
+pub fn get_baseline_offset(metrics: &FontMetrics, mode:Baseline) -> f32 {
   match mode{
-    Baseline::Top => -metrics.ascent as f64,
-    Baseline::Hanging => metrics.cap_height as f64,
-    Baseline::Middle => metrics.cap_height as f64 / 2.0,
+    Baseline::Top => -metrics.ascent,
+    Baseline::Hanging => metrics.cap_height,
+    Baseline::Middle => metrics.cap_height / 2.0,
     Baseline::Alphabetic => 0.0,
-    Baseline::Ideographic => -metrics.descent as f64,
-    Baseline::Bottom => -metrics.descent as f64,
+    Baseline::Ideographic => -metrics.descent,
+    Baseline::Bottom => -metrics.descent,
   }
 }
